requirements.txt
openai==1.37.1
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.111.1
uvicorn[standard]==0.30.4
motor[srv]==3.5.1
python-jose==3.3.0
passlib==1.7.4
